Updated Login.tsx and App.tsx
